+++ components/layout/main-layout.tsx
@@ -5,7 +5,7 @@ import type React from "react"
 import { useState } from "react"
 import Link from "next/link"
 import { usePathname } from "next/navigation"
-import { Button, Card } from "@/ui"
+import { Button, Card } from "@/components/ui"
 import { LayoutDashboard, Smartphone, MessageSquare, Settings, Activity, Menu, X, LogOut, User, Contact } from "lucide-react"
 import { cn } from "@/lib/utils"
 import { Notifications } from "@/components/notifications"
