+++ components/layout/sidebar.tsx
@@ -4,7 +4,7 @@ import { useState } from "react"
 import Link from "next/link"
 import { usePathname } from "next/navigation"
 import { cn } from "@/lib/utils"
-import { Button, Badge, Separator } from "@/ui"
+import { Button, Badge, Separator } from "@/components/ui"
 import {
   LayoutDashboard,
   Smartphone,
