--- ui/index.ts
@@ -1,8 +1,0 @@
-export * from "./button"
-export * from "./card"
-export * from "./input"
-export * from "./select"
-export * from "./badge"
-export * from "./dropdown-menu"
-export * from "./avatar"
-export * from "./separator"
